third_party/WebKit/LayoutTests/html/dialog/synthetic-click-inert.html

<!DOCTYPE html>
<html>
<style>
dialog {
    /* Don't needlessly span the whole screen. */
    width: 50px;
}
</style>
<body>
<script src="../../resources/js-test.js"></script>
<button>Click me</button>
<div id="div">Click me too</div>
<dialog></dialog>
<script>
description('Test that inert nodes still get programmatic click events');
dialog = document.querySelector('dialog');
dialog.showModal();

button = document.querySelector('button');
div = document.getElementById('div');
clicked = null;

[button, div].forEach(function(element) {
    element.addEventListener('click', function(e) { clicked = element; });
    expectedElement = element;

    clicked = null;
    debug('Calling click() on ' + element.tagName);
    element.click();
    shouldBe('clicked', 'expectedElement');

    clicked = null;
    debug('Calling dispatchEvent() on ' + element.tagName);
    element.dispatchEvent(new Event('click'));
    shouldBe('clicked', 'expectedElement');
});
</script>
</body>
</html>
