third_party/WebKit/LayoutTests/html/dialog/synthetic-click-inert-expected.txt

Test that inert nodes still get programmatic click events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Calling click() on BUTTON
PASS clicked is expectedElement
Calling dispatchEvent() on BUTTON
PASS clicked is expectedElement
Calling click() on DIV
PASS clicked is expectedElement
Calling dispatchEvent() on DIV
PASS clicked is expectedElement
PASS successfullyParsed is true

TEST COMPLETE
Click me
Click me too
